chore: moving views to views and out of component
